+++ packages/driver-flow/src/FlowDriver.ts
@@ -122,11 +122,11 @@ export default class FlowDriver extends Driver<FlowConfig> {
 
   // https://github.com/facebook/flow/blob/e466b0ee519622a8977e89708be156a73e570ef0/hack/utils/exit_status.ml#L78
   // https://github.com/facebook/flow/blob/e466b0ee519622a8977e89708be156a73e570ef0/src/common/flowExitStatus.ml#L54
-  handleFailure(error: Execution) {
+  processFailure(error: Execution) {
     if (error.code === 2) {
       this.tool.logError(error.stdout); // Command failures
     } else {
-      super.handleFailure(error);
+      super.processFailure(error);
     }
   }
 }